Rename docs examples to be more reasonable
