+++ docs/components/guide-examples/custom/point-cloud.tsx
